Tilemap and improved food spawning
Added a tilemap to make the environment more interesting to move around in, there is also a separate layer of tilemap that restricts the player to a certain area.

Added a check for if a food prefab is outside of the bounds that the player is in. This means that the player will not have unreachable food taking up the food cap if they more close to the edge of the map.