New Food Type and fixed a tilemap bug.
A new "largefood" object was added that functions similarly to the regular food object, it just spawns slightly slower, when score is implemented it will give a higher score than the normal food.

There was also an annoying layering bug with the bound tiles which was fixed by setting its layer order lower, which revealed the tiles that were clashing with the bound tiles. I then removed these and the bound tiles now render correctly.

I also added back the food despawn code which somehow disappeared in the last version.