Additions of comments, enemy and several misc changes to polish the game.
Added an enemy that will look at, and follow the player constantly. There is no way to kill this enemy and they will keep multiplying, and adding difficult to the game. This gives the game an actual objective, which is to survive for as long as possible and avoid touching the enemy.

Added a death scene to the canvas object. This triggers when the player object collides with an enemy object. This presents the player with the option to respawn, which reloads the scene.

Code is now fully commented to the extent of what I understand.